fix(cooldown): store absolute expiry instead of remaining duration
ActiveCooldowns stored cooldown as remaining seconds, decremented by
DeltaTime each TickActiveCooldowns call. Under GMC client prediction
this drifted: combined client moves cause GenAncillaryTick to fire
multiple times per real frame on the same predicted move, with the
current move's full (accumulating) DeltaTime each time. Observed
~2-2.6x faster than wall clock on a laggy client, draining a 5s
cooldown in ~2s.

Switch storage to absolute expiry time in ActionTimer units (the
same scheme effect EndAtActionTimer already uses). SetCooldownForAbility
computes ActionTimer + CooldownTime once; GetCooldownForAbility returns
max(0, Expiry - ActionTimer); TickActiveCooldowns becomes a GC pass.
Repeated tick fires on the same move are idempotent because expiry is
a fixed point in time, not a counter.

Public API surface unchanged — SetCooldownForAbility still takes a
duration, GetCooldownForAbility still returns remaining seconds.
Internal storage type widened to double to match ActionTimer.

+++ Source/GMCAbilitySystem/Public/Components/GMCAbilityComponent.h
@@ -977,8 +977,17 @@ private:
 	UPROPERTY()
 	TMap<int, UGMCAbility*> ActiveAbilities;
 	
+	// Active cooldowns keyed by ability tag. Value is the absolute expiry
+	// time in ActionTimer units (i.e. ActionTimer at which the cooldown
+	// ends). Storing as expiry-time rather than remaining-duration is
+	// required to be drift-free under GMC client prediction: combined
+	// client moves cause AncillaryTick to fire multiple times per real
+	// frame on the same move, and a remaining-duration model accumulates
+	// extra decrements (observed 2-2.6x faster than wall clock on a
+	// laggy client). Expiry-time is set once and only compared, so
+	// repeated tick fires are harmless.
 	UPROPERTY()
-	TMap<FGameplayTag, float> ActiveCooldowns;
+	TMap<FGameplayTag, double> ActiveCooldowns;
 
 	int GenerateAbilityID() const {return ActionTimer * 100;}
